initial work on getting undefined command working
