+++ bin/tapestry
@@ -1,0 +1,57 @@
+#!/usr/bin/env node
+
+const program = require('commander')
+const exec = require('child_process').exec
+const path = require('path')
+const webpack = require('webpack')
+
+const pkg = require('../package.json')
+
+const cwdCached = process.cwd()
+
+program
+  .version(pkg.version)
+  .command('build')
+  .action(function() {
+    process.chdir(path.join(__dirname, '..'))
+    exec('babel ./src --out-dir ./dist --presets=es2015,react', (err, stdout, stderr) => {
+      if (err) {
+        console.error(err)
+        return;
+      }
+      console.log(stdout)
+      console.log('Built Server')
+      const webpackConfig = require('../dist/webpack.config.js').default
+      const compiler = new webpack(webpackConfig(cwdCached, true))
+      // No feedback coming out of this yet
+      compiler.run((err, stats) => {
+        const jsonStats = stats.toJson()
+        if (err)
+          console.error(err)
+        if (jsonStats.errors.length > 0)
+          console.error('Errors: ', jsonStats.errors)
+        if (jsonStats.warnings.length > 0)
+          console.warn('Warnings: ', jsonStats.warnings)
+      })
+    })
+  })
+
+program
+  .command('start')
+  .action(function() {
+    global.TAPESTRY_PRODUCTION = true
+    console.log(global.TAPESTRY_PRODUCTION)
+    require('../dist/boot.js')
+  })
+
+program.parse(process.argv)
+
+var NO_COMMAND_SPECIFIED = program.args.length === 0;
+
+if (NO_COMMAND_SPECIFIED) {
+  require('babel-register')
+  require('../src/boot.js')
+}
+
+
+
