+++ src/webpack.config.js
@@ -1,14 +1,15 @@
 import path from 'path'
 import webpackCleanPlugin from 'clean-webpack-plugin'
 
-export default (context) => ({
+export default (context, isTree = false) => ({
   resolve: {
     modulesDirectories: ['node_modules', context, `${context}/node_modules`]
   },
-  entry: 'tapestry/src/client.js',
+  entry: isTree ? `${context}/tapestry.js` : 'tapestry-wp/src/client.js',
   output: {
-    path: path.resolve(context, 'public'),
-    filename: 'bundle.js'
+    path: path.resolve(context, isTree ? 'dist' : 'public'),
+    filename: isTree ? 'tree.js' : 'bundle.js',
+    libraryTarget: isTree ? 'commonjs2' : 'var'
   },
   module: {
     loaders: [{
@@ -17,6 +18,16 @@ export default (context) => ({
       loader: 'babel'
     }]
   },
+  externals: [
+    {
+      'isomorphic-fetch': {
+        root: 'isomorphic-fetch',
+        commonjs2: 'isomorphic-fetch',
+        commonjs: 'isomorphic-fetch',
+        amd: 'isomorphic-fetch'
+      }
+    }
+  ],
   plugins: [
     new webpackCleanPlugin(['public'], { root: context, verbose: false })
     // new webpack.optimize.UglifyJsPlugin(),
